+++ visuals/src/main/kotlin/io/redgreen/timelapse/visuals/Commit.kt
@@ -2,5 +2,5 @@ package io.redgreen.timelapse.visuals
 
 data class Commit(
   val insertions: Int,
-  val deletions: Int = 0
+  val deletions: Int
 )